Create README.md
Designing a Grocery application using django web framework Using python programming language and javascript with Signup registration, Login functions, and Logout functions with the help of frontend works HTML, CSS, and Bootstrap. In this application, I developed to order groceries using the app And we can add the groceries to the cart, and we can view our orders in the app.